requirements.txt
PyPDF2
pdfplumber
openai
python-frontmatter
tqdm
unidecode